Refactor diff_tables.TableDiffer -> hashdiff_tables.HashDiffer
