+++ tests/common.py
@@ -43,7 +43,8 @@ if os.environ.get("LOG_LEVEL", False):
     level = getattr(logging, os.environ["LOG_LEVEL"].upper())
 
 logging.basicConfig(level=level)
-logging.getLogger("diff_tables").setLevel(level)
+logging.getLogger("hashdiff_tables").setLevel(level)
+logging.getLogger("joindiff_tables").setLevel(level)
 logging.getLogger("table_segment").setLevel(level)
 logging.getLogger("database").setLevel(level)
